repos/MM-ACT/requirements.txt
transformers==4.46.0
wandb
webdataset
omegaconf
pyarrow
pandas
jaxtyping
diffusers==0.32.2
typeguard
accelerate
deepspeed
lightning
datasets
image-reward
torchmetrics==1.6.2
trl==0.17.0
math_verify 
gradio>=4.44.1
lerobot==0.3.2
scipy==1.15.3
h5py==3.14.0
transforms3d==0.4.2
